Moved xml-init such that global-code can work.
In the original version xml-init (and thus) global code was ran before the xml-file was installed. If the global code executed code of an action the initialization was triggered again and did fail.
Now relocated xml-init such that the file exists in cache already. It's the users reponsibility to only call code in the global section that does not need uninitialized code from the global section.