Added routine to do a generic chuncked write.
Via a binary search the data is writen to database. If one
row fails this specific row will be omitted. A binary write is fast
when only a limited number of rows fail as it tries to write big blocks.

Note when a now() field is part of the key the chunked write might even
resolve a key-conflicts as different batches have a slightly different
time of insertion (with TIMESTAMPS, not with DATE-TYPES).